Working colour gradients in minimap

+++ CSC8503/MapNode.cpp
@@ -33,5 +33,5 @@ float MapNode::getColourRatio() {
 		redCount += i->redTally;
 		blueCount += i->blueTally;
 	}
-	return (redCount - blueCount) / nodeTally;
+	return (float)(redCount - blueCount) / nodeTally;
 }
